etl_pipeline/requirements.txt
dagster==1.7.2
dagit==1.7.2
dagster-postgres
pandas==1.5.3
pyarrow==16.0.0
minio==7.2.5
pymysql
numpy==1.23.5
cryptography==42.0.5
psycopg2-binary==2.9.9
dagster-dbt == 0.23.2
pyspark==3.3.2
dbt-core==1.5
dbt-postgres==1.5
bs4
pytz==2022.7.1